requirements_docker.txt
numpy
pandas
requests
praw
transformers 
torch
twikit
bs4
alpaca-py

// openai



// #btw if u would like to use this on docker, just click f1 and select dev container open folder in container and open the .devcontainer 

// u use *python "filename.py"* to run the file in the terminal eg python trading_bot.py